+++ BasicShooterAssignment/Assets/__Main/Script/Main/ScriptableObject/CharacterStats/ICharacterStats.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public interface ICharacterStats 
 {
-  int CharacterDamageAbsolute { get; }
-  int WeaponDamageAbsolute { get; }
-  int CharacterUpgradePercentageModifier { get; }
-  int AttackDamageAbsoluteModifier { get; }
-  int AttackDamagePercentModifier { get; }
+  float CharacterDamageAbsolute { get; }
+  float WeaponDamageAbsolute { get; }
+  float CharacterUpgradePercentageModifier { get; }
+  float AttackDamageAbsoluteModifier { get; }
+  float AttackDamagePercentModifier { get; }
 }
